pyproject.toml
[tool.poetry]
name = "user-management-micro-service-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
aioboto3 = "^11.3.0"
aiohttp = "^3.8.6"
black = "^23.9.1"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
aioredis = "^2.0.1"
pyjwt = "^2.8.0"
isort = "^5.12.0"
uvicorn = "^0.23.2"
passlib = "^1.7.4"
pydantic-settings = "^2.0.3"
pip = "^23.2.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
sqlalchemy-utils = "^0.41.1"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
